refactor(police): extract walkShadowRoots helper
`getShadowAwareSelection` and `findHostRoot` rolled their own
descend-through-shadow-roots walk — same stack-and-visit shape, two
copies. Extract `walkShadowRoots(host, visit)` and route both callers
through it.

`nodeInside` stays as-is — it walks the parent chain ascending and hops
shadow hosts on the way up, a different shape that the descend-visitor
helper doesn't fit.
